+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/CommandGroups/ArmPositions/ExampleCommandGroup.java
@@ -1,13 +1,13 @@
-package org.firstinspires.ftc.teamcode.CommandGroups;
+package org.firstinspires.ftc.teamcode.CommandGroups.ArmPositions;
 
 import com.arcrobotics.ftclib.command.InstantCommand;
 import com.arcrobotics.ftclib.command.SequentialCommandGroup;
 import com.arcrobotics.ftclib.geometry.Pose2d;
 import com.arcrobotics.ftclib.geometry.Rotation2d;
 import com.arcrobotics.ftclib.geometry.Translation2d;
-
 import org.firstinspires.ftc.teamcode.Commands.FollowPath;
 import org.firstinspires.ftc.teamcode.RobotContainer;
+
 
 import java.util.ArrayList;
 
@@ -17,123 +17,88 @@ import java.util.ArrayList;
 // ParallelRaceGroup
 // ParallelDeadlineGroup
 
-public class SweepAlliancePieces extends SequentialCommandGroup {
+public class ExampleCommandGroup extends SequentialCommandGroup {
 
     // constructor
-    public SweepAlliancePieces() {
+    public ExampleCommandGroup() {
 
         addCommands (
-                new InstantCommand(()-> RobotContainer.odometry.setCurrentPos(new Pose2d(-0.40,1.6,new Rotation2d(Math.toRadians(-90))))),
+                new InstantCommand(()-> RobotContainer.odometry.setCurrentPos(new Pose2d(0.9,1.6,new Rotation2d(Math.toRadians(-90))))),
 
                 new FollowPath(
-                        1.0,
-                        1.0,
-                        0.0,
-                        0.0,
-                        new Rotation2d(Math.toRadians(-90.0)),
-                        new ArrayList<Translation2d>() {{ }},
-                        new Pose2d(-0.85, 1.1, new Rotation2d(Math.toRadians(-90.0))),
-                        new Rotation2d(Math.toRadians(-90.0))
-                ),
-
-                new FollowPath(
-                        1.0,
-                        1.0,
-                        0.0,
-                        0.0,
-                        new Rotation2d(Math.toRadians(-90.0)),
-                        new ArrayList<Translation2d>() {{ }},
-                        new Pose2d(-0.85, 0.3, new Rotation2d(Math.toRadians(-90.0))),
-                        new Rotation2d(Math.toRadians(-90.0))
-                ),
-                new FollowPath(
-                        1.0,
-                        1.0,
-                        0.0,
-                        0.0,
-                        new Rotation2d(Math.toRadians(180.0)),
-                        new ArrayList<Translation2d>() {{ }},
-                        new Pose2d(-1.15, 0.3, new Rotation2d(Math.toRadians(180.0))),
-                        new Rotation2d(Math.toRadians(-90.0))
-                ),
-                new FollowPath(
-                        1.0,
+                        2.0,
                         1.0,
                         0.0,
                         0.0,
                         new Rotation2d(Math.toRadians(90.0)),
                         new ArrayList<Translation2d>() {{ }},
-                        new Pose2d(-1.15, 1.35, new Rotation2d(Math.toRadians(90.0))),
-                        new Rotation2d(Math.toRadians(-90.0))
+                        new Pose2d(1.45, 1.45, new Rotation2d(Math.toRadians(45.0))),
+                        new Rotation2d(Math.toRadians(-135.0))
                 ),
 
                 new FollowPath(
-                        1.0,
+                        2.0,
                         1.0,
                         0.0,
                         0.0,
-                        new Rotation2d(Math.toRadians(-90.0)),
+                        new Rotation2d(Math.toRadians(-135.0)),
                         new ArrayList<Translation2d>() {{ }},
-                        new Pose2d(-1.15, 0.3, new Rotation2d(Math.toRadians(-90.0))),
+                        new Pose2d(-1.24, 0.95, new Rotation2d(Math.toRadians(-90.0))),
                         new Rotation2d(Math.toRadians(-90.0))
                 ),
 
                 new FollowPath(
-                        1.0,
-                        1.0,
-                        0.0,
-                        0.0,
-                        new Rotation2d(Math.toRadians(180.0)),
-                        new ArrayList<Translation2d>() {{ }},
-                        new Pose2d(-1.45, 0.3, new Rotation2d(Math.toRadians(180.0))),
-                        new Rotation2d(Math.toRadians(-90.0))
-
-                ),
-
-                new FollowPath(
-                        1.0,
+                        2.0,
                         1.0,
                         0.0,
                         0.0,
                         new Rotation2d(Math.toRadians(90.0)),
                         new ArrayList<Translation2d>() {{ }},
-                        new Pose2d(-1.45, 1.35, new Rotation2d(Math.toRadians(90.0))),
-                        new Rotation2d(Math.toRadians(-90.0))
+                        new Pose2d(1.45, 1.45, new Rotation2d(Math.toRadians(45.0))),
+                        new Rotation2d(Math.toRadians(-135.0))
                 ),
+
                 new FollowPath(
-                        1.0,
+                        2.0,
                         1.0,
                         0.0,
                         0.0,
-                        new Rotation2d(Math.toRadians(-90.0)),
+                        new Rotation2d(Math.toRadians(-135.0)),
                         new ArrayList<Translation2d>() {{ }},
-                        new Pose2d(-1.45, 0.3, new Rotation2d(Math.toRadians(-90.0))),
+                        new Pose2d(-1.50, 0.95, new Rotation2d(Math.toRadians(-90.0))),
                         new Rotation2d(Math.toRadians(-90.0))
                 ),
+
                 new FollowPath(
-                        1.0,
-                        1.0,
-                        0.0,
-                        0.0,
-                        new Rotation2d(Math.toRadians(180.0)),
-                        new ArrayList<Translation2d>() {{ }},
-                        new Pose2d(-1.59, 0.3, new Rotation2d(Math.toRadians(180.0))),
-                        new Rotation2d(Math.toRadians(-90.0))
-                ),
-                new FollowPath(
-                        1.0,
+                        2.0,
                         1.0,
                         0.0,
                         0.0,
                         new Rotation2d(Math.toRadians(90.0)),
                         new ArrayList<Translation2d>() {{ }},
-                        new Pose2d(-1.59, 1.35, new Rotation2d(Math.toRadians(90.0))),
+                        new Pose2d(1.45, 1.45, new Rotation2d(Math.toRadians(45.0))),
+                        new Rotation2d(Math.toRadians(-135.0))
+                ),
+                new FollowPath(
+                        2.0,
+                        1.0,
+                        0.0,
+                        0.0,
+                        new Rotation2d(Math.toRadians(-135.0)),
+                        new ArrayList<Translation2d>() {{ }},
+                        new Pose2d(-1.75, 0.95, new Rotation2d(Math.toRadians(-90.0))),
                         new Rotation2d(Math.toRadians(-90.0))
+                ),
+                new FollowPath(
+                        2.0,
+                        1.0,
+                        0.0,
+                        0.0,
+                        new Rotation2d(Math.toRadians(90.0)),
+                        new ArrayList<Translation2d>() {{ }},
+                        new Pose2d(1.45, 1.45, new Rotation2d(Math.toRadians(45.0))),
+                        new Rotation2d(Math.toRadians(-135.0))
                 )
-
-
-
-
 
         // new command1
         // new command2
