+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/CommandGroups/ArmPositions/HuntingPos.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.CommandGroups;
+package org.firstinspires.ftc.teamcode.CommandGroups.ArmPositions;
 
 import com.arcrobotics.ftclib.command.InstantCommand;
 import com.arcrobotics.ftclib.command.SequentialCommandGroup;
@@ -11,29 +11,31 @@ import org.firstinspires.ftc.teamcode.RobotContainer;
 // ParallelRaceGroup
 // ParallelDeadlineGroup
 
-public class SpecimenPlacePos extends SequentialCommandGroup {
+public class HuntingPos extends SequentialCommandGroup {
 
     // constructor
-    public SpecimenPlacePos() {
+    public HuntingPos() {
+        addCommands(
 
-        addCommands (
-        // What this position should do is give the camera a good vantage point as well as keep the arm out of the way
 
-                // lifts the shoulder up 90+-60 degrees
-                // lifts the shoulder up to 135 degrees
-                new InstantCommand(() ->RobotContainer.shoulderJoint.RotateTo(135)),
+                // sets the elbow to a straight position at 135 degrees
+                new InstantCommand(() -> RobotContainer.shoulderJoint.RotateTo(130)),
 
-                // folds the elbow in 270
-                new InstantCommand(() ->RobotContainer.elbowJoint.RotateTo(255)),
+                // sets elbow to a straight position at 145 degrees,
+                // 145 is becouse gravity and leverage bend the elbow down so the extera 10 degrees stops the drooping
+                new InstantCommand(() ->RobotContainer.elbowJoint.RotateTo(145)),
 
-                // folds the wrist in 0
-                new InstantCommand(() -> RobotContainer.flappyFlappyWrist.RotateTo(15)),
+                // sets wrist to a straight position
+                new InstantCommand(() -> RobotContainer.flappyFlappyWrist.RotateTo(45)),
 
-                // powers the wrist and moves it to straight position
+                // powers and sets wrist to a straight position
                 new InstantCommand(() -> RobotContainer.wristRotateServo.RotateTo(180))
 
 
+
         );
+
+
     }
 
 }
