+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/CommandGroups/ArmPositions/StartingArmStowHigh.java
@@ -1,8 +1,9 @@
-package org.firstinspires.ftc.teamcode.CommandGroups;
+package org.firstinspires.ftc.teamcode.CommandGroups.ArmPositions;
 
 import com.arcrobotics.ftclib.command.InstantCommand;
 import com.arcrobotics.ftclib.command.SequentialCommandGroup;
 
+import org.firstinspires.ftc.teamcode.Commands.Pause;
 import org.firstinspires.ftc.teamcode.RobotContainer;
 
 // Example Sequential Command Group
@@ -11,24 +12,30 @@ import org.firstinspires.ftc.teamcode.RobotContainer;
 // ParallelRaceGroup
 // ParallelDeadlineGroup
 
-public class BackDepositePose extends SequentialCommandGroup {
+public class StartingArmStowHigh extends SequentialCommandGroup {
 
     // constructor
-    public BackDepositePose() {
+    public StartingArmStowHigh() {
 
         addCommands (
+        // What this position should do is give the camera a good vantage point as well as keep the arm out of the way
 
-                // lifts the shoulder up 45 degrees
-                new InstantCommand(() ->RobotContainer.shoulderJoint.RotateTo(45)),
+                //powers shoulder
+                new InstantCommand(() ->RobotContainer.shoulderJoint.RotateTo(60)),
+                // folds the elbow in 270
+                new InstantCommand(() ->RobotContainer.elbowJoint.RotateTo(270)),
 
-                // folds the elbow in 60 degrees
-                new InstantCommand(() ->RobotContainer.elbowJoint.RotateTo(195)),
+                new Pause(1.5),
+                // lifts the shoulder up 90+-60 degrees
+                // lifts the shoulder up to 135 degrees
+                new InstantCommand(() ->RobotContainer.shoulderJoint.RotateTo(135)),
 
-                // folds the wrist in 135 degrees
-                new InstantCommand(() -> RobotContainer.flappyFlappyWrist.RotateTo(170)),
+                // folds the wrist in 0
+                new InstantCommand(() -> RobotContainer.flappyFlappyWrist.RotateTo(0)),
 
-                // folds the wrist in 135 degrees
+                // powers the wrist and moves it to straight position
                 new InstantCommand(() -> RobotContainer.wristRotateServo.RotateTo(180))
+
 
 
         );
