+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/CommandGroups/AutomatedMovements/BlueSideHighBucketDeposit.java
@@ -1,10 +1,20 @@
-package org.firstinspires.ftc.teamcode.CommandGroups;
+package org.firstinspires.ftc.teamcode.CommandGroups.AutomatedMovements;
 
 import com.arcrobotics.ftclib.command.InstantCommand;
 import com.arcrobotics.ftclib.command.SequentialCommandGroup;
+import com.arcrobotics.ftclib.geometry.Pose2d;
+import com.arcrobotics.ftclib.geometry.Rotation2d;
+import com.arcrobotics.ftclib.geometry.Translation2d;
 
+import org.firstinspires.ftc.teamcode.CommandGroups.ArmPositions.ArmStowHigh;
+import org.firstinspires.ftc.teamcode.CommandGroups.ArmPositions.BackDepositePose;
+import org.firstinspires.ftc.teamcode.Commands.FollowPath;
+import org.firstinspires.ftc.teamcode.Commands.OpenClaw;
 import org.firstinspires.ftc.teamcode.Commands.Pause;
 import org.firstinspires.ftc.teamcode.RobotContainer;
+import org.firstinspires.ftc.teamcode.Subsystems.SlideTargetHeight;
+
+import java.util.ArrayList;
 
 // Example Sequential Command Group
 // There are also:
@@ -12,32 +22,39 @@ import org.firstinspires.ftc.teamcode.RobotContainer;
 // ParallelRaceGroup
 // ParallelDeadlineGroup
 
-public class StartingArmStowHigh extends SequentialCommandGroup {
+public class BlueSideHighBucketDeposit extends SequentialCommandGroup {
 
     // constructor
-    public StartingArmStowHigh() {
+    public BlueSideHighBucketDeposit() {
 
         addCommands (
-        // What this position should do is give the camera a good vantage point as well as keep the arm out of the way
+                new FollowPath(
+                        1.0,
+                        1.0,
+                        0.0,
+                        0.0,
+                        new Rotation2d(Math.toRadians(-90.0)),
+                        new ArrayList<Translation2d>() {{ }},
+                        new Pose2d(1.53, 1.43, new Rotation2d(Math.toRadians(-135))),
+                        new Rotation2d(Math.toRadians(-135))),
 
-                //powers shoulder
-                new InstantCommand(() ->RobotContainer.shoulderJoint.RotateTo(60)),
-                // folds the elbow in 270
-                new InstantCommand(() ->RobotContainer.elbowJoint.RotateTo(270)),
+                new Pause(1),
 
-                new Pause(1.5),
-                // lifts the shoulder up 90+-60 degrees
-                // lifts the shoulder up to 135 degrees
-                new InstantCommand(() ->RobotContainer.shoulderJoint.RotateTo(135)),
+                new InstantCommand(()-> RobotContainer.linearSlide.moveTo(SlideTargetHeight.SAMPLE_HIGH)),
 
-                // folds the wrist in 0
-                new InstantCommand(() -> RobotContainer.flappyFlappyWrist.RotateTo(0)),
+                new Pause(1),
 
-                // powers the wrist and moves it to straight position
-                new InstantCommand(() -> RobotContainer.wristRotateServo.RotateTo(180))
+                new BackDepositePose(),
 
+                new Pause(1),
 
+                new OpenClaw(),
 
+                new Pause(1),
+
+                new ArmStowHigh(),
+
+                new InstantCommand(()-> RobotContainer.linearSlide.moveTo(SlideTargetHeight.SAMPLE_ZERO))
         );
     }
