+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/CommandGroups/AutomatedMovements/PlaceSpecimenAddOffset.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.CommandGroups;
+package org.firstinspires.ftc.teamcode.CommandGroups.AutomatedMovements;
 
 import com.arcrobotics.ftclib.command.InstantCommand;
 import com.arcrobotics.ftclib.command.SequentialCommandGroup;
@@ -6,12 +6,11 @@ import com.arcrobotics.ftclib.geometry.Pose2d;
 import com.arcrobotics.ftclib.geometry.Rotation2d;
 import com.arcrobotics.ftclib.geometry.Translation2d;
 
-import org.firstinspires.ftc.teamcode.Commands.CloseClaw;
-import org.firstinspires.ftc.teamcode.Commands.FollowPath;
+import org.firstinspires.ftc.teamcode.CommandGroups.ArmPositions.SpecimenPlacePos;
+import org.firstinspires.ftc.teamcode.Commands.GoToNextDropOff;
 import org.firstinspires.ftc.teamcode.Commands.OpenClaw;
 import org.firstinspires.ftc.teamcode.Commands.Pause;
 import org.firstinspires.ftc.teamcode.RobotContainer;
-import org.firstinspires.ftc.teamcode.Subsystems.PivotingWrist;
 import org.firstinspires.ftc.teamcode.Subsystems.SlideTargetHeight;
 
 import java.util.ArrayList;
@@ -22,107 +21,46 @@ import java.util.ArrayList;
 // ParallelRaceGroup
 // ParallelDeadlineGroup
 
-public class GroundCyclingAuto extends SequentialCommandGroup {
+public class PlaceSpecimenAddOffset extends SequentialCommandGroup {
 
     // constructor
-    public GroundCyclingAuto() {
+    public PlaceSpecimenAddOffset() {
 
         addCommands (
-                // pick up middle
-                new FollowPath(
+
+                //sets arm to specimen place position
+                new SpecimenPlacePos(),
+                //sets the slides to low
+                new InstantCommand(()-> RobotContainer.linearSlide.moveTo(SlideTargetHeight.SAMPLE_LOW)),
+
+                new Pause(1),
+                //max speed = 1
+                new GoToNextDropOff(
                         2.0,
-                        1.0,
+                        1.2,
                         0.0,
                         0.0,
-                        new Rotation2d(Math.toRadians(-135)),
+                        new Rotation2d(Math.toRadians(-90.0)),
                         new ArrayList<Translation2d>() {{ }},
-                        new Pose2d(1.56, 1.05, new Rotation2d(Math.toRadians(-90))),
-                        new Rotation2d(Math.toRadians(-90.0))),
+                        new Pose2d(0.25, 0.805, new Rotation2d(Math.toRadians(-90.0))),
+                        new Rotation2d(Math.toRadians(-90))),
 
-                new Pause(0.5),
+                new Pause(1),
 
-                new HuntingPos(),
+                new InstantCommand(()-> RobotContainer.linearSlide.moveTo(SlideTargetHeight.SAMLE_SPECIMEN)),
 
-                new Pause(0.5),
+                new Pause(1),
 
-                new DropToGrab(),
+                new OpenClaw(),
 
-                new Pause(0.5),
+                new Pause(0.25),
 
-                new CloseClaw(),
+                new InstantCommand(()-> RobotContainer.linearSlide.moveTo(SlideTargetHeight.SAMPLE_ZERO)),
 
-                new Pause(0.5),
-
-                new ArmStowHigh(),
-
-                new BlueSideHighBucketDeposit(),
-
-                // pick up right
-                new FollowPath(
-                        2.0,
-                        1.0,
-                        0.0,
-                        0.0,
-                        new Rotation2d(Math.toRadians(-135)),
-                        new ArrayList<Translation2d>() {{ }},
-                        new Pose2d(1.31, 1.05, new Rotation2d(Math.toRadians(-90))),
-                        new Rotation2d(Math.toRadians(-90.0))),
-
-                new Pause(0.5),
-
-                new HuntingPos(),
-
-                new Pause(0.5),
-
-                new DropToGrab(),
-
-                new Pause(0.5),
-
-                new CloseClaw(),
-
-                new Pause(0.5),
-
-                new ArmStowHigh(),
-
-                new BlueSideHighBucketDeposit()
-
-//                //pick up left
-//                new FollowPath(
-//                        1.0,
-//                        1.0,
-//                        0.0,
-//                        0.0,
-//                        new Rotation2d(Math.toRadians(-135)),
-//                        new ArrayList<Translation2d>() {{ }},
-//                        new Pose2d(1.51, 0.97, new Rotation2d(Math.toRadians(-45))),
-//                        new Rotation2d(Math.toRadians(-45))),
-//
-//                new Pause(2),
-//
-//                new InstantCommand(() -> RobotContainer.wristRotateServo.RotateTo(180)),
-//
-//                new HuntingPos(),
-//
-//                new Pause(1),
-//
-//                new DropToGrab(),
-//
-//                new Pause(1),
-//
-//                new CloseClaw(),
-//
-//                new Pause(1),
-//
-//                new ArmStowHigh(),
-//
-//                new BlueSideHighBucketDeposit()
-
-                // x - 1.209
-                // y - 1.3655
-                //
-
-
+                new Pause(1)
         );
+
+
     }
 
 }
