+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/teleop/RedTeleop.java
@@ -3,12 +3,8 @@ package org.firstinspires.ftc.teamcode.opmode.teleop;
 import com.arcrobotics.ftclib.command.CommandOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.firstinspires.ftc.teamcode.CommandGroups.ArmStowHigh;
-import org.firstinspires.ftc.teamcode.CommandGroups.ArmStowLow;
-import org.firstinspires.ftc.teamcode.CommandGroups.ExampleCommandGroup;
-import org.firstinspires.ftc.teamcode.CommandGroups.StartingArmStowHigh;
+import org.firstinspires.ftc.teamcode.CommandGroups.ArmPositions.StartingArmStowHigh;
 import org.firstinspires.ftc.teamcode.Commands.CloseClaw;
-import org.firstinspires.ftc.teamcode.Commands.OpenClaw;
 import org.firstinspires.ftc.teamcode.RobotContainer;
 
 /*
